+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.1.4
+- Created `array.iterate()` to iterate an `Array` using an `Iterator`
+- Created `array.iterator()` to return a `Generator` to iterate an `Array`, requires native Generators / from meta programming (no transpiling)
+
 ## 1.1.3
 - Moved more `variables` to `constants`
 - Another pass of default parameter values